acquire lock for reading database as well
This is required. I this the reason is when system writes the file it
first append to it then moves the index to start of it forward. If these
two operations happen at the same time it can lead to reading content of
previous file plus new content that is being written to it altogether
which results in invalid JSONs.